Fold the variational base class into basehmm
